Move CLI path setup policy into Domain
