+++ Assets/Tests/Editor/CliPathSetupWriterTests.cs
@@ -4,7 +4,7 @@ using System.IO;
 
 using NUnit.Framework;
 
-using io.github.hatayama.UnityCliLoop.Application;
+using io.github.hatayama.UnityCliLoop.Domain;
 using io.github.hatayama.UnityCliLoop.Infrastructure;
 
 namespace io.github.hatayama.UnityCliLoop.Tests.Editor
