+++ Packages/src/Editor/Domain/CliPathSetupProfileResolver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0582561513a84d43a0d418430a2f4b3c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
